fix rounding error bug finding second face
When a body with a selected face is tilted in X,Y or Z,
the area of equally sized faces may be off by an
insignificant amount due to the precision of doubles.
This caused a failure when finding the second face.
The plugin now rounds the area for each face to three
decimal points, to mimic what is displayed by F360 in the GUI.