initial support for unequal faces
If the parallel face is a different size,
due to an angled adjacent face or something
similar, the edge used for the secondary axis
of the rectangular pattern might not be found.
If this happens, the plugin will create a sketch
line in 3D to use as the secondary axis.